+++ Editor/NDMF/OptimizingPlugin.cs
@@ -8,6 +8,7 @@ using UnityEngine;
 using VRC.SDK3.Avatars.Components;
 using VRC.SDK3.Avatars.ScriptableObjects;
 using Object = UnityEngine.Object;
+using static VRChatExpressionParametersOptimizer.Editor.RecursiveWalker;
 
 [assembly: ExportsPlugin(typeof(VRChatExpressionParametersOptimizer.Editor.NDMF.OptimizingPlugin))]
 namespace VRChatExpressionParametersOptimizer.Editor.NDMF
@@ -68,48 +69,6 @@ namespace VRChatExpressionParametersOptimizer.Editor.NDMF
             ObjectRegistry.RegisterReplacedObject(lhs, source);
             lhs = source;
         }
-        
-        private static int CountBits(IEnumerable<VRCExpressionParameters.Parameter> parameters) =>
-            parameters.Aggregate(0, (bits, param) => bits + VRCExpressionParameters.TypeCost(param.valueType));
-        
-        private static IEnumerable<AnimatorTransitionBase> GatherTransitions(AnimatorStateMachine asm)
-        {
-            var transitions = new List<IEnumerable<AnimatorTransitionBase>> { asm.entryTransitions, asm.anyStateTransitions, };
-            Debug.Log($"VEPO - {asm.name}: checking {asm.stateMachines.Length} state machines");
-            if (asm.stateMachines.Length > 0)
-            {
-                var op = asm.stateMachines.Select(y =>
-                    GatherTransitions(y.stateMachine));
-                transitions.AddRange(op);
-            }
-            else
-            {
-                transitions.AddRange(asm.stateMachines.Select(y => asm.GetStateMachineTransitions(y.stateMachine)));
-                transitions.AddRange(asm.states.Select(x => x.state).Select(x => x.transitions));
-            }
-                    
-            Debug.Log($"VEPO - {asm.name}: checking {transitions.Count} transitions");
-
-            return transitions.SelectMany(op => op);
-        }
-
-        private static IEnumerable<VRCExpressionsMenu.Control.Parameter> ReferencedParametersFromMenu(VRCAvatarDescriptor ad)
-        {
-            return Walk(ad.expressionsMenu);
-
-            IEnumerable<VRCExpressionsMenu.Control.Parameter> Walk(VRCExpressionsMenu menu)
-            {
-                return menu.controls.SelectMany(x => new[] { x.parameter }
-                    .Chain(x.subParameters ?? Array.Empty<VRCExpressionsMenu.Control.Parameter>())
-                    .Chain(x.subMenu != null ? Walk(x.subMenu) : Array.Empty<VRCExpressionsMenu.Control.Parameter>()));
-            }
-        }
-    }
-
-    internal static class LinqExtensions
-    {
-        internal static IEnumerable<T> Chain<T>(this IEnumerable<T> self, IEnumerable<T> other) 
-            => new[] { self, other }.SelectMany(a => a);
     }
 }
 #endif
