+++ Editor/NDMF/VRChatExpressionParametersOptimizer.Editor.NDMF.asmdef
@@ -4,7 +4,8 @@
   "allowUnsafeCode": false,
   "rootNamespace": "",
   "references": [
-    "GUID:62ced99b048af7f4d8dfe4bed8373d76"
+    "GUID:62ced99b048af7f4d8dfe4bed8373d76",
+    "GUID:5bfc3880dc46439991c16a8cd706664c"
   ],
   "versionDefines": [
     {
